refactored to have Post.js
